Subcomments show up correctly
