+++ app/templates/deal_dashboard.html
@@ -45,6 +45,7 @@
                                     </div>
                                 </div>
                                 <!-- Voting Column -->
+                                {% if current_user.is_authenticated %}
                                 <div class="col-1 d-flex flex-column align-items-center justify-content-center">
                                     <button type="button" class="btn btn-link p-0 mb-0" onclick="upvote('{{ deal['deal_id'] }}')">
                                         <b>▲</b>
@@ -54,6 +55,19 @@
                                         <b>▼</b>
                                     </button>
                                 </div>
+                                {% else %}
+                                <div class="text-center">
+                                    <a href="{{ url_for('auth.login') }}" class="btn btn-link p-0 mb-0">
+                                        <b>🔒</b>
+                                    </a>
+                                    <br> <!-- New line for separation -->
+                                    <span id="vote-count-{{ deal['deal_id'] }}">{{ deal.get('upvotes', 0) - deal.get('downvotes', 0) }}</span>
+                                    <br> <!-- New line for separation -->
+                                    <a href="{{ url_for('auth.login') }}" class="btn btn-link p-0 mt-0">
+                                        <b>🔒</b>
+                                    </a>
+                                </div>
+                                {% endif %}
                             </div>
                         </div>
                     </div>
@@ -139,5 +153,6 @@
     </div>
 
     <script src="{{ url_for('static', filename='js/index.js') }}"></script>
+    <script src="{{ url_for('static', filename='js/view_comments.js') }}"></script>
     <script src="{{ url_for('static', filename='js/deal_details.js') }}"></script>
 {% endblock %}
